added GCN
The GCN still has the bug, couldnt run.
The bug should be in the model, which fails to convolute the features.
(returning with this error:
UserWarning: Using a target size (torch.Size([1, 1])) that is different to the input size (torch.Size([1, 235, 1])). This will likely lead to incorrect results due to broadcasting. Please ensure they have the same size.
)